# api/openapi.yaml
openapi: 3.1.0
info:
  title: Universal Gateway (Jev Decision + Multi-LLM Orchestration)
  version: 0.1.0
  description: |
    北向统一入口。/v1/execute 为主入口（LLM 编排 + Jev 决策 + 门控执行）；
    /v1/decide 为 Jev 直通校验版（服务端持 key）；/v1/chat/completions 为 OpenAI 兼容透传。
servers:
  - url: http://localhost:8080
paths:
...
  /admin/policies:
    get:
      summary: 查看阈值（需 X-Admin-Token）
      responses:
        '200':
          description: 阈值
          content:
            application/json:
              schema:
                type: object
    put:
      summary: 在线更新阈值（热生效，带校验；需 X-Admin-Token）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 生效结果
          content:
            application/json:
              schema:
                type: object
  /admin/jev:
    patch:
      summary: 更新 Jev endpoint/model_pin/timeout/retry（需 X-Admin-Token）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 生效结果
          content:
            application/json:
              schema:
                type: object
  /admin/jev/key:
    post:
      summary: 轮换 Jev API key，即时生效（需 X-Admin-Token）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 生效结果（key 脱敏）
          content:
            application/json:
              schema:
                type: object
  /admin/llm/providers:
    get:
      summary: 列出 LLM providers（需 X-Admin-Token）
      responses:
        '200':
          description: provider 列表
          content:
            application/json:
              schema:
                type: object
    post:
      summary: 新增或更新 provider（需 X-Admin-Token）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 生效结果
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: 删除 provider（需 X-Admin-Token；默认 provider 需先切换）
      responses:
        '200':
          description: 生效结果
          content:
            application/json:
              schema:
                type: object
  /admin/llm/keys:
    post:
      summary: 轮换某 provider key（需 X-Admin-Token）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 生效结果
          content:
            application/json:
              schema:
                type: object
  /admin/keys:
    get:
      summary: 列出租户 keys（name + 脱敏 key + 创建时间；需 X-Admin-Token）
      responses:
        '200':
          description: 租户列表
          content:
            application/json:
              schema:
                type: object
    post:
      summary: 给租户签发 key（{name, key?}，不带 key 则生成；明文只返回一次；需 X-Admin-Token）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 新 key 明文（仅此一次）
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: 吊销租户 key（按 ?name= 或 ?key=；需 X-Admin-Token）
      responses:
        '200':
          description: 生效结果
          content:
            application/json:
              schema:
                type: object
...